+++ tests/testthat/test-apps.R
@@ -1,0 +1,7 @@
+testthat::skip_on_cran()
+
+library(shinytest2)
+
+lapply(list.files("apps", full.names = TRUE), function(folder) {
+  shinytest2::test_app(folder)
+})