+++ NEWS.md
@@ -6,6 +6,7 @@
 - Added `shinyjs.version` JavaScript variable to help troubleshooting
 - Replace `digest::digest()` with `rlang::hash`, to reduce dependencies and it's faster (#248)
 - Replace `list(...)` with `rlang::list2(...)` which is more flexible and allows you to use trailing commas without error
+- Added tests using {shinytest2}
 
 # shinyjs 2.1.0 (2021-12-20)